Error Handling done

+++ library/src/main/java/com/library/library/service/BookStudentService.java
@@ -1,9 +1,12 @@
 package com.library.library.service;
 
 import com.library.library.dao.BookStudentDAO;
-import com.library.library.dao.OverlapDAO;
+import com.library.library.dao.StudentDAO;
 import com.library.library.entity.BookStudent;
 import com.library.library.entity.Request;
+import com.library.library.entity.Student;
+import com.library.library.exception.BookStudentNotFound;
+import com.library.library.exception.StudentNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,16 +16,21 @@ import java.util.List;
 public class BookStudentService {
 
     private BookStudentDAO bookStudentDAO;
-    private OverlapDAO overlapDAO;
+    private StudentDAO studentDAO;
 
     @Autowired
-    public BookStudentService (BookStudentDAO bookStudentDAO, OverlapDAO overlapDAO){
+    public BookStudentService (BookStudentDAO bookStudentDAO,StudentDAO studentDAO){
         this.bookStudentDAO = bookStudentDAO;
-        this.overlapDAO = overlapDAO;
+        this.studentDAO = studentDAO;
     }
 
-    public BookStudent getBookStudentById(Integer id){
-        return bookStudentDAO.findBookStudentById(id);
+    public BookStudent getBookStudentById(Integer id) throws BookStudentNotFound {
+        BookStudent bs = bookStudentDAO.findBookStudentById(id);
+        if(bs==null){
+            throw new BookStudentNotFound();
+        }else {
+            return bs;
+        }
     }
 
     public List<BookStudent> getBookStudentByStudentId(Integer id){
@@ -37,16 +45,25 @@ public class BookStudentService {
         return bookStudentDAO.findAllBookStudent();
     }
 
-    public BookStudent addNewBookStudentPair(BookStudent bookStudent){
+    public BookStudent addNewBookStudentPair(BookStudent bookStudent)throws StudentNotFoundException{
+        Integer st = bookStudent.getStudent().getId();
+        Student student1 = this.studentDAO.getStudentById(st);
+        if(student1==null) {
+            throw new StudentNotFoundException();
+        }
         return bookStudentDAO.makeBookStudent(bookStudent);
     }
 
-    public void deleteById(Integer id){
-        bookStudentDAO.deleteBookStudentById(id);
+    public void deleteById(Integer id) throws BookStudentNotFound{
+        BookStudent bs = bookStudentDAO.findBookStudentById(id);
+        if(bs==null){
+            throw new BookStudentNotFound();
+        }else {
+            bookStudentDAO.deleteBookStudentById(id);
+        }
     }
 
     public Boolean doesRequestOverlap(Request request){
-        List<BookStudent> bs = this.overlapDAO.overlap(request);
-        return !bs.isEmpty();
+        return this.bookStudentDAO.checkOverlap(request);
     }
 }
